requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
aerich==0.7.2
    # via -r requirements.in
aiosqlite==0.20.0
    # via tortoise-orm
annotated-types==0.7.0
    # via pydantic
anyio==4.6.2.post1
    # via starlette
asyncpg==0.30.0
    # via -r requirements.in
beautifulsoup4==4.12.3
    # via -r requirements.in
certifi==2024.8.30
    # via requests
charset-normalizer==3.4.0
    # via requests
click==8.1.7
    # via
    #   aerich
    #   uvicorn
dictdiffer==0.9.0
    # via aerich
fastapi==0.115.4
    # via -r requirements.in
h11==0.14.0
    # via uvicorn
idna==3.10
    # via
    #   anyio
    #   requests
iso8601==2.1.0
    # via tortoise-orm
pydantic==2.9.2
    # via
    #   -r requirements.in
    #   aerich
    #   fastapi
    #   tortoise-orm
pydantic-core==2.23.4
    # via pydantic
pypika-tortoise==0.2.1
    # via tortoise-orm
pytz==2024.2
    # via tortoise-orm
requests==2.32.3
    # via -r requirements.in
sniffio==1.3.1
    # via anyio
soupsieve==2.6
    # via beautifulsoup4
starlette==0.41.2
    # via fastapi
tomlkit==0.13.2
    # via aerich
tortoise-orm==0.21.7
    # via
    #   -r requirements.in
    #   aerich
typing-extensions==4.12.2
    # via
    #   aiosqlite
    #   fastapi
    #   pydantic
    #   pydantic-core
urllib3==2.2.3
    # via requests
uvicorn==0.32.0
    # via -r requirements.in
